+++ src/backend/index.ts
@@ -1,5 +1,5 @@
 // Adonis plugin:
-export * from '../interfaces.js';
+export * from '../types.js';
 export { configure } from './configure.js';
 export { stubsRoot } from './stubs/main.js';
 export { defineConfig } from './define_config.js';
