+++ src/frontend/fields/helpers/message-test.vue
@@ -42,8 +42,8 @@
 <script setup lang="ts">
 import AppLayout from '../../shared/app-layout.vue';
 import ContentHeader from '../../shared/content-header.vue';
-import type { SharedPageProps } from '../../../interfaces';
-import { ResponseStatus } from '../../../interfaces';
+import type { SharedPageProps } from '../../../types';
+import { ResponseStatus } from '../../../types';
 import { useSharedStore } from '../../store';
 
 const props = defineProps<SharedPageProps>();
