+++ src/frontend/fields/interface/ui-page.vue
@@ -49,8 +49,8 @@ import AppLayout from '../../shared/app-layout.vue';
 import StickyHeader from '../../shared/sticky-header.vue';
 import { router } from '@inertiajs/vue3';
 
-import { ResponseStatus } from '../../../interfaces';
-import type { UiItem, UiPageProps, SharedPageProps } from '../../../interfaces';
+import { ResponseStatus } from '../../../types';
+import type { UiItem, UiPageProps, SharedPageProps } from '../../../types';
 import UiCard from '../interface/components/ui-card.vue';
 
 type ModelType = { [key: string]: string | undefined };
