+++ src/frontend/fields/null-field.vue
@@ -7,7 +7,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue';
-import type { FieldSpec } from '../../interfaces';
+import type { FieldSpec } from '../../types';
 import { commonProps } from '../shared/helpers';
 const props = defineProps({
   ...commonProps,
