+++ src/frontend/fields/pages/page-meta-box.vue
@@ -26,7 +26,7 @@
 </template>
 <script setup lang="ts">
 import { computed } from 'vue';
-import type { PageMeta } from '../../../interfaces';
+import type { PageMeta } from '../../../types';
 import { formatDate } from '../../shared/helpers';
 import Icon from '../../shared/icon.vue';
