+++ src/frontend/fields/pages/pages-index.story.vue
@@ -17,7 +17,7 @@
 <script setup lang="ts">
 import PagesIndex from './pages-index.vue';
 import { sharedProps } from '../../../frontend/test/mocks';
-import type { PageItem } from '../../../interfaces';
+import type { PageItem } from '../../../types';
 
 const pages: PageItem[] = [
   {
