+++ src/frontend/fields/string-field.vue
@@ -34,7 +34,7 @@
 
 <script setup lang="ts">
 import { computed, ref, nextTick } from 'vue';
-import type { FieldSpec } from '../../interfaces';
+import type { FieldSpec } from '../../types';
 import { useModelStore, useSharedStore } from '../store/index';
 import { commonProps } from '../shared/helpers';
