+++ src/frontend/shared/helpers.ts
@@ -1,5 +1,5 @@
 import { App, PropType } from 'vue';
-import { FieldSpec } from '../../interfaces';
+import { FieldSpec } from '../../types';
 import { BibleBooksMap } from './bibleBooks';
 import { Variant, Story } from 'histoire';
