+++ src/frontend/shared/index-filter.story.vue
@@ -23,7 +23,7 @@
 
 <script setup lang="ts">
 import IndexFilter from './index-filter.vue';
-import type { TabItem, PageItem } from '../../interfaces';
+import type { TabItem, PageItem } from '../../types';
 import { ref, computed } from 'vue';
 import PageIndexItem from '../fields/pages/page-index-item.vue';
